Line tracking program
This is the most simple line tracking program.
The trategy here is to trace the center of the sensor(s) that detect(s)
the white line. If no sensor has detected the white line, the robot will
move according to last status. The speed of the servo is about
60degree/0.06s.
Each loop takes about 30ms. This means if each loop we only change
30degree, the servo motor will finish the motion before next loop. Here
the maximum change is set to 20 degree.