Line sensor: PD control + 90 degree
This is a simple implementation of PD + 90 degree turn.
1 new sensor is added in to the system, which could dectect whether an
object appears within a specific distance. The sensor is put on the left
of the robot car. and it points downwards.
The program goes through several modes:
START=>SEARCH=>TURN=>ZIGZAG
(START MODE)At the beginning, the newly added sensor could detect the
ground. As the robot car climb higher and higher, the sensor will no
longer detect the ground. (SEARCH MODE) Then we move to next mode to
search for the bridge. When the robot car move to the 90 degree turn.
The sensor could detected the bridge. (TURN MODE) After that, we disable
the sensors on the right. (ZIGZAG MODE)After the robot have turned the
90 degree corner, the newly added sensor is disabled.